video: Made ruffle_video crate, moved core video Error to it
