+++ core/Cargo.toml
@@ -16,6 +16,7 @@ indexmap = "1.9.1"
 log = "0.4"
 minimp3 = { version = "0.5.1", optional = true }
 ruffle_render = { path = "../render" }
+ruffle_video = { path = "../video" }
 ruffle_macros = { path = "macros" }
 ruffle_wstr = { path = "../wstr" }
 swf = { path = "../swf" }
