+++ core/src/backend/video.rs
@@ -3,35 +3,14 @@
 use generational_arena::{Arena, Index};
 use ruffle_render::backend::RenderBackend;
 use ruffle_render::bitmap::BitmapInfo;
+use ruffle_video::error::Error;
 use swf::{VideoCodec, VideoDeblocking};
-use thiserror::Error;
 
 mod software;
 
 pub use crate::backend::video::software::SoftwareVideoBackend;
 
 pub type VideoStreamHandle = Index;
-
-#[derive(Error, Debug)]
-pub enum Error {
-    #[error("Attempted to seek to omitted frame {0} without prior decoded frame")]
-    SeekingBeforeDecoding(u32),
-
-    #[error("Unsupported video codec type: {0:?}")]
-    UnsupportedCodec(VideoCodec),
-
-    #[error("Video stream is not registered")]
-    VideoStreamIsNotRegistered,
-
-    #[error("Couldn't create bitmap for video frame")]
-    BitmapError(#[from] ruffle_render::error::Error),
-
-    #[error("Video decoding isn't supported")]
-    DecodingNotSupported,
-
-    #[error(transparent)]
-    DecoderError(Box<dyn std::error::Error + Send + Sync>),
-}
 
 /// An encoded video frame of some video codec.
 pub struct EncodedFrame<'a> {
