+++ core/src/backend/video/software/decoders/vp6.rs
@@ -1,5 +1,6 @@
 use crate::backend::video::software::VideoDecoder;
-use crate::backend::video::{DecodedFrame, EncodedFrame, Error, FrameDependency};
+use crate::backend::video::{DecodedFrame, EncodedFrame, FrameDependency};
+use ruffle_video::error::Error;
 
 use h263_rs_yuv::bt601::yuv420_to_rgba;
